Adding AppNavigator
